% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AALatlas}
\alias{AALatlas}
\title{Automated Anatomical Labeling (AAL) atlas coordinates}
\format{
A dataframe with 90 rows and 4 columns/variables.
}
\source{
\doi{10.1006/nimg.2001.0978}
}
\usage{
AALatlas
}
\description{
A dataframe of the Automated Anatomical Labeling (AAL) atlas from the work of Tzourio-Mazoyer et al. (2002) atlas to use with the \code{\link{net.3dplot}()}
function. Each row corresponds to a region of interest (ROI) to be plotted using the Montreal Neurological Institute (MNI) space. The first
column corresponds to the community labels (in this atlas, there are none, therefore this column is filled with NA), and the second, third, and
fourth columns correspond to the X, Y, and Z coordinates of the ROIs in MNI space, respectively. See Tzourio-Mazoyer et al. (2002) <doi:10.1006/nimg.2001.0978>
for more details.
}
\keyword{datasets}
