% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AALfmri}
\alias{AALfmri}
\title{90 ROI data from the NYU test-retest resting state fMRI data set}
\format{
A data matrix with 197 rows and 90 columns/variables, where each column corresponds to an ROI from the AAL atlas.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
AALfmri
}
\description{
A data matrix of the second scan from the first subject from the NYU test-restest resting-state fMRI data set.
}
\keyword{datasets}
