% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjmatrix}
\alias{adjmatrix}
\title{Adjacency matrix of fMRI data set}
\format{
A matrix with 333 rows and 333 columns, where each entry denotes whether two nodes in the Gordon atlas are clustered together.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
adjmatrix
}
\description{
The adjacency matrix calculated from the "fmridata" data set, using the Gordon atlas.
}
\keyword{datasets}
