% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{data_setup}
\usage{
data_setup(data, n.subj, T)
}
\arguments{
\item{data}{Multivariate time series, \eqn{T}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}

\item{n.subj}{Total number of subjects in the dataset}

\item{T}{Number of time points in the dataset for each subject}
}
\value{
List of original data, \eqn{T}, where each element in the list is a subject and their associated multivariate time series
}
\description{
This function is embedded in the main FaBiSearch function. It serves to take the original input time series, \eqn{T}, and conver it into a more convenient
list format. Note, this assumes that timecourse length for each subject is the same.
}
