% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.network.R
\name{est.net}
\alias{est.net}
\title{est.net}
\usage{
est.net(data, nruns = 50, lambda = 7, rank = "optimal", algtype = "brunet")
}
\arguments{
\item{data}{Multivariate time series, \eqn{Y}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}

\item{nruns}{Number of runs to use for NMF function, by default is set to =50}

\item{lambda}{Either a value >1 to use hierarchical clustering on the consensus matrix, or a between 0 and 1 inclusive to
use a cutoff value for the consensus matrix}

\item{rank}{Specifies whether to find rank using optimization procedure (specify "optimal") or use a predetermined rank (integer value, e.g., 4),
by default is set to ="optimal"}

\item{algtype}{Type of algorithm for NMF function, by default is set to ="brunet"}
}
\value{
Adjacency matrix noting temporal dependencies between variables from \eqn{Y}
}
\description{
Compliment function to FaBiSearch. Estimates sparse stationary networks using non-negative matrix factorization (NMF).
}
