% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_network.R
\name{estimate_network}
\alias{estimate_network}
\title{estimate_network}
\usage{
estimate_network(data, alg.type, n.rank, n.runs, method, lambda)
}
\arguments{
\item{data}{Multivariate time series, \eqn{T}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}

\item{alg.type}{Type of algorithm for NMF function -> check ?nmf for details, under "method"}

\item{n.rank}{Rank value to be used in NMF calculations}

\item{n.runs}{Number of runs to use for NMF function}

\item{method}{Either "cluster" to use hierarchical clustering on the consensus matrix, or "cutoff" to use a cutoff value for the consensus matrix}

\item{lambda}{Threshold parameter, relates to number of clusters if method = "cluster", or a value between 0 and 1 (inclusive) if method = "cutoff"}
}
\value{
Adjacency matrix noting temporal dependencies between variables from \eqn{T}
}
\description{
Compliment function to FaBiSearch. Estimates sparse stationary networks using non-negative matrix factorization (NMF).
}
