% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmridata}
\alias{fmridata}
\title{fMRI data set}
\format{
A matrix with 35 rows and 333 columns/variables, where each column corresponds to an ROI from the Gordon atlas.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
fmridata
}
\description{
A data set of the first subject and their corresponding first stationary block from the second fMRI scan.
}
\keyword{datasets}
