% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gordatlas}
\alias{gordatlas}
\title{Gordon atlas coordinates}
\format{
A dataframe with 333 rows and 4 columns/variables.
}
\source{
\doi{10.1093/cercor/bhu239}
}
\usage{
gordatlas
}
\description{
A dataframe of the Gordon et al. (2016) atlas to use with the \code{\link{net.3dplot}()} function. Each row corresponds to a region of interest
(ROI) to be plotted using the Montreal Neurological Institute (MNI) space. The first column corresponds to the community labels as a string,
and the second, third, and fourth columns correspond to the X, Y, and Z coordinates of the ROIs in MNI space, respectively. See Gordon et al. (2016)
<doi:10.1093/cercor/bhu239> for more details.
}
\keyword{datasets}
