% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gordfmri}
\alias{gordfmri}
\title{333 ROI data from the NYU test-retest resting state fMRI data set}
\format{
A data matrix with 197 rows and 333 columns/variables, where each column corresponds to an ROI from the Gordon atlas.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
gordfmri
}
\description{
A data matrix of the second scan from the first subject from the NYU test-restest resting-state fMRI data set.
}
\keyword{datasets}
