% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logSP500}
\alias{logSP500}
\title{Daily adjusted logarithmic returns for the Standard and Poor's 500}
\format{
A dataframe with 815 rows and 500 columns/variables.
}
\usage{
logSP500
}
\description{
A dataframe of the daily adjusted logarithmic returns for the Standard and Poor's 500 (S&P 500) stock market index. Each row corresponds to a
trading day from 2018-01-01 to 2021-03-31. Data was retrieved from Yahoo Finance using the \code{getSymbols()} function from the \code{quantmod}
package.
}
\keyword{datasets}
