% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.rank.R
\name{opt.rank}
\alias{opt.rank}
\title{opt.rank}
\usage{
opt.rank(data, nruns = 50, algtype = "brunet")
}
\arguments{
\item{data}{A matrix with time points in rows and variables in columns. This input is the multivariate time series, \eqn{Y}, to be analyzed.}

\item{nruns}{A positive integer, by default is set to 5. Defines the number of runs to use for NMF function.}

\item{algtype}{A character string denoting the type of algorithm for NMF function, by default is set to "brunet".}
}
\value{
Integer denoting the optimal rank found
}
\description{
This function finds the optimal rank to use on a dataset using non-negative matrix factorization
}
\examples{
## Finding the optimal rank for an input dataset "data" with the default settings
opt.rank(data)
[1] 5

## Finding the optimal rank for an input dataset "data" with nruns of 100 and the default "brunet" algorithm
opt.rank(data, nruns = 100)
[1] 4

## Finding the optimal rank for an input dataset "data" using the least square NMF method and the default nruns
opt.rank(data, algtype = "ls-nmf")
[1] 8

}
\references{
"Factorized Binary Search: a novel technique for change point detection in multivariate high-dimensional time series networks", Ondrus et al
(2021), preprint.
}
\author{
Martin Ondrus, \email{mondrus@ualberta.ca}, Ivor Cribben, \email{cribben@ualberta.ca}
}
