% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_split.R
\name{permute_split}
\alias{permute_split}
\title{permute_split}
\usage{
permute_split(data)
}
\arguments{
\item{data}{Multivariate time series, \eqn{Y}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}
}
\value{
Permuted version of the original multivariate time series, \eqn{Y}
}
\description{
This function is embedded in the main FaBiSearch function. It serves to permute the input data by whole rows, that is, permute the time course of all ROIs
together
}
