% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.network.R
\name{plot.net}
\alias{plot.net}
\title{plot.net}
\usage{
\method{plot}{net}(adj.matrix, communities = NULL, colors = NULL)
}
\arguments{
\item{adj.matrix}{A 333 by 333 adjacency matrix}

\item{communities}{A vector of strings related to communities to plot - by default, all communities are plotted - (any combination of "Default", "SMhand", "SMmouth", "Visual", "FrontoParietal", "Auditory", "None", "CinguloParietal", "RetrosplenialTemporal", "CinguloOperc", "VentralAttn", "Salience", "DorsalAttn)}

\item{colors}{A vector of hex codes for node colors to distinguish each community - by default, each community is given a predefined color}
}
\value{
a 3D plot of the estimated stationary network from the adjacency matrix
}
\description{
This function uses a Gordon atlas defined adjacency matrix and returns a 3D plot of the estimated stationary network of this adjacency matrix
}
