% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_splits.R
\name{refit_splits}
\alias{refit_splits}
\title{refit_splits}
\usage{
refit_splits(orig.splits, curr.subj, T, x, n.rep, n.rank, alg.type)
}
\arguments{
\item{alg.type}{}
}
\value{

}
\description{
refit_splits
}
