% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim2}
\alias{sim2}
\title{Simulation 2 data set}
\format{
A matrix with 200 rows and 80 columns/variables.
}
\usage{
sim2
}
\description{
A data set of Simulation 2, which is generated from a multivariate Gaussian distribution with 2 clusters where the correlation between nodes in the same
cluster is 0.75 and between different clusters is 0.2. This data set contains a change point at t=100 where vertex labels are randomly reshuffled.
}
\keyword{datasets}
