% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AALfmri}
\alias{AALfmri}
\title{90 ROI data from the NYU test-retest resting state fMRI data set}
\format{
An object of class \code{data.frame} with 197 rows and 90 columns.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
AALfmri
}
\description{
A data matrix of the second scan from the first subject from the NYU test-retest resting-state fMRI data set.
}
\details{
A data matrix of the second scan from the first subject from the NYU test-retest resting-state fMRI data set. Variables/nodes are defined
using the Tzourio-Mazoyer et al. (2002) AAL atlas, which can be accessed as the \code{AALatlas} dataframe.
}
\keyword{datasets}
