% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjmatrix}
\alias{adjmatrix}
\title{Adjacency matrix for the NYU test-restest resting-state fMRI data set}
\format{
A 333 * 333 matrix, where each entry takes a value 1 (0) if two nodes are (not) connected by an edge, using the Gordon atlas.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt/}
}
\usage{
adjmatrix
}
\description{
The adjacency matrix calculated from the \code{gordfmri} data set, using the Gordon atlas.
}
\keyword{datasets}
