% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.cps.R
\name{detect.cps}
\alias{detect.cps}
\title{detect.cps}
\usage{
detect.cps(
  data,
  mindist = 35,
  nruns = 50,
  nreps = 100,
  alpha = "p-value",
  rank = "optimal",
  algtype = "brunet"
)
}
\arguments{
\item{data}{Multivariate time series, \eqn{Y}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}

\item{mindist}{Minimum distance between change points, by default is set to =35}

\item{nruns}{Number of runs to use for NMF function, by default is set to =50}

\item{nreps}{Number of repetitions for bootstrapping procedure, by default is set to =100}

\item{alpha}{Significance level cutoff for deciding whether to retain change points or not, by default is set to ="p-value" and returns the associated
p-value of the p-value test}

\item{rank}{Specifies whether to find rank using optimization procedure (specify "optimal") or use a predetermined rank (integer value, e.g., 4),
by default is set to ="optimal"}

\item{algtype}{Type of algorithm for NMF function, by default is set to ="brunet"}
}
\value{
Outputs a list where:\cr
\code{$rank} is the rank used for change point detection \cr
\code{$change.points} is a table of the change points detected where column "T" is the time of the change point and "stat.test" is the result of the t-test\cr
\code{$compute.time} is the compute time for the algorithm\cr
}
\description{
This function takes a multivariate time series, \eqn{Y}, and returns the change points detected. Utilizes non-negative
factorization (NMF) to detect changes in clustering structure.
}
