% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.net.R
\name{est.net}
\alias{est.net}
\title{est.net}
\usage{
est.net(data, nruns = 50, lambda = 7, rank = "optimal", algtype = "brunet")
}
\arguments{
\item{data}{A matrix with time points in rows and variables in columns. This input is the multivariate time series, \eqn{Y}, to be analyzed.}

\item{nruns}{A positive integer, by default is set to 5. Defines the number of runs to use for NMF function.}

\item{lambda}{Either a positive integer to use hierarchical clustering on the consensus matrix, or a positive real number to use as a cutoff value for
the consensus matrix.}

\item{rank}{A positive integer denoting the value of rank to use in the algorithm. By default is set to "optimal" so the function finds the optimal rank
to use.}

\item{algtype}{A character string denoting the type of algorithm for NMF function, by default is set to "brunet".}
}
\value{
Adjacency matrix noting temporal dependencies between variables from \eqn{Y}
}
\description{
Compliment function to FaBiSearch. Estimates sparse stationary networks using non-negative matrix factorization (NMF). Each run of NMF
produces an adjacency matrix, and by averaging across all adjacency matrices we generate a consensus matrix where each entry is the probability of two
variables or nodes being clustered together
}
\examples{
## Estimating the network for a multivariate dataset, "data" using default settings - outputs as an adjacency matrix
est.net(data)

## Estimating the network for a multivariate dataset, "data", specifying the number of runs to 100
## and using hierarchical clustering to generate the adjacency matrix with a cutoff value of 7 clusters
est.net(data, nruns = 100, lambda = 7)

## Estimating the network for a multivariate dataset, "data", specifying the number of runs to 100
## and using a cutoff value for the adjacency matrix to enforce sparsity, where the cutoff is 0.5
est.net(data, nruns = 100, lambda = 0.5)

## Estimating the network for a multivariate dataset, "data", specifying the rank beforehand at 4
est.net(data, rank = 4)

## Estimating the network for a multivariate dataset, "data", using the least square NMF method
est.net(data, algtype = "ls-nmf")

}
\references{
"Factorized Binary Search: a novel technique for change point detection in multivariate high-dimensional time series networks", Ondrus et al
(2021), preprint.
}
\author{
Martin Ondrus, \email{mondrus@ualberta.ca}, Ivor Cribben, \email{cribben@ualberta.ca}
}
