% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaBiSearch.R
\name{FaBiSearch}
\alias{FaBiSearch}
\title{FaBiSearch}
\usage{
FaBiSearch(
  output.name,
  data,
  which.subj,
  n.subj = NULL,
  T = NULL,
  min.dist,
  n.rep,
  alpha,
  method.rank,
  n.runs,
  alg.type,
  test.type
)
}
\arguments{
\item{test.type}{}
}
\value{

}
\description{
FaBiSearch
}
