% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmridata}
\alias{fmridata}
\title{Data from the NYU test-retest resting state fMRI data set}
\format{
A data matrix with 35 rows and 333 columns/variables, where each column corresponds to an ROI from the Gordon atlas.
}
\source{
\url{https://www.nitrc.org/projects/nyu_trt}
}
\usage{
fmridata
}
\description{
A data matrix from the first stationary block of the second scan from the first subject from the NYU test-restest resting-state fMRI data set.
}
\keyword{datasets}
