% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.rank.R
\name{opt.rank}
\alias{opt.rank}
\title{Finds the optimal rank for non-negative matrix factorization (NMF)}
\usage{
opt.rank(Y, nruns = 50, algtype = "brunet")
}
\arguments{
\item{Y}{A numerical matrix representing the multivariate time series, with the columns representing its components.}

\item{nruns}{A positive integer with default value equal to 50. It is used to define the number of runs in the NMF function.}

\item{algtype}{A character string, which defines the algorithm to be used in the NMF function. By default it is set to "brunet". See the "Algorithms" section of
\code{\link[NMF]{nmf}} for more information on the available algorithms.}
}
\value{
A positive integer representing the optimal rank.
}
\description{
This function finds the optimal rank for non-negative matrix factorization (NMF).
}
\examples{
## Finding the optimal rank for an input data set "sim2" with the default settings
\donttest{opt.rank(sim2)}
# [1] 3

## Finding the optimal rank for an input data set "sim2" with nruns of 100 and the
## default "brunet" algorithm
\donttest{opt.rank(sim2, nruns = 100)}
# [1] 3

## Finding the optimal rank for an input data set "sim2" using the "snmf/l"
## NMF method and the default nruns
\donttest{opt.rank(sim2, algtype = "snmf/l")}
# [1] 4

}
\references{
"Factorized Binary Search: a novel technique for change point detection in multivariate high-dimensional time series networks", Ondrus et al.
(2021), preprint.
}
\author{
Martin Ondrus, \email{mondrus@ualberta.ca}, Ivor Cribben, \email{cribben@ualberta.ca}
}
