% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.rank.R
\name{opt.rank}
\alias{opt.rank}
\title{opt.rank}
\usage{
opt.rank(data, nruns = 50, algtype = "brunet")
}
\arguments{
\item{data}{Multivariate time series, \eqn{Y}, to be analyzed, should be in a matrix format with time points in rows and variables in columns}

\item{nruns}{Number of runs to use for NMF function, by default is set to =50}

\item{algtype}{Type of algorithm for NMF function, by default is set to ="brunet"}
}
\value{
Integer denoting the optimal rank found
}
\description{
This function finds the optimal rank to use on a dataset using non-negative matrix factorization
}
