% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_splits.R
\name{refit_splits}
\alias{refit_splits}
\title{refit_splits}
\usage{
refit_splits(orig.splits, curr.subj, T, x, n.rep, n.rank, alg.type)
}
\arguments{
\item{orig.splits}{Matrix of candidate change points (\code{T.split}) and the change in loss (\code{chg.loss})}

\item{curr.subj}{Multivariate time series data from the current subject}

\item{x}{Time series of data}

\item{n.rep}{Number of repetitions for bootstrapping procedure}

\item{n.rank}{Rank value for NMF function}

\item{alg.type}{Type of algorithm for NMF function -> check ?nmf for details, under "method"}
}
\value{
List where each element is a unique change point and contains a vector of loss values from the refitted NMF calculations
}
\description{
This function is embedded in the main FaBiSearch function. It serves to recalculate the loss at each candidate change point considering the fact that
multiple change points might have been detected.
}
