% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_splits.R
\name{sign_splits}
\alias{sign_splits}
\title{sign_splits}
\usage{
sign_splits(orig.splits, refit.splits, perm.distr, alpha, test.type)
}
\arguments{
\item{test.type}{}
}
\value{

}
\description{
sign_splits
}
