% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_all.R
\name{split_all}
\alias{split_all}
\title{split_all}
\usage{
split_all(
  data,
  split.index,
  lower,
  upper,
  x,
  min.dist,
  n.runs,
  n.rank,
  alg.type
)
}
\arguments{
\item{alg.type}{}
}
\value{

}
\description{
split_all
}
